Add io-ts support + single file config
This commit adds support for generating io-ts configs that can also
output the same types that the interfaces would provide.

Generating io-ts files can be done by calling the "generateIoTs" task
in a project.

Support for using different renderers for different files has also
been added trough the "codegen-output" attribute on the "<form>"
element in an XP xml-file.

The recommended approach is to use the "generateTypeScript" task for
most files, but add "codegen-output='IoTs'"  to the interfaces you
need io-ts validation on (user input).

It is also possible to set "codegen-output='None'"  so that output is
not generated at all.

Closes #22, closes #25